+++ src/util/DragaliaActionsUtils.js
@@ -2,9 +2,9 @@ import { useDispatch } from 'react-redux';
 import { useContext } from 'react';
 import { useSelector } from 'react-redux';
 import { MappingContext } from '../components/SaveEditor';
-import { addJsonDataListObject, replaceJsonDataListObject, addToObjectListObjectField
-
- } from '../actions/JsonDataActions';
+import { addJsonDataListObject, replaceJsonDataListObject, addToObjectListObjectField,
+    updateJsonDataObjectField 
+} from '../actions/JsonDataActions';
 
 import { WeaponBuildupPieceType } from '../enum/Enums';
 
@@ -331,6 +331,45 @@ const useDragaliaActions = () => {
         }
     }
 
+    const maxTutorial = () => {
+        const tutorialFlagList = [
+            1001,
+            1002,
+            1003,
+            1004,
+            1005,
+            1006,
+            1007,
+            1008,
+            1009,
+            1010,
+            1011,
+            1012,
+            1013,
+            1014,
+            1015,
+            1016,
+            1017,
+            1018,
+            1019,
+            1020,
+            1021,
+            1022,
+            1023,
+            1024,
+            1025,
+            1026,
+            1027,
+            1028,
+            1029,
+            1030
+        ];
+        const tutorialStatus = 60999;
+
+        dispatch(updateJsonDataObjectField("user_data", "tutorial_flag_list", tutorialFlagList));
+        dispatch(updateJsonDataObjectField("user_data", "tutorial_status", tutorialStatus));
+    }
+
     return { 
         addAdventurerStory,
         addDragonStory,
@@ -339,7 +378,8 @@ const useDragaliaActions = () => {
         addDragon,
         addWeaponSkin,
         handleWeaponBuildupSkins,
-        handleWeaponBuildupSkinsAll
+        handleWeaponBuildupSkinsAll,
+        maxTutorial
     };
 };